introduce KinSyl (for hiquae'qurim) and accented i's; related TODOs
The tasks in README.md relate to working with KinSyl and {Rim}Syl text
on Mac and Linux (GNOME and swaywm) dev environments. Apologies to
Windows users, I am not really a Windows'hnym, so taking time out of
my schedule to support Windows or even file accurate bug reports is
not on the cards. KeyMan does support Windows, though, so typing text
should eventually work ok for you!

(The linked EJB Cree typewriter layout is almost ready for KinSyl, but
it lacks the Inuktitut qaai{/quae} row that's pretty essential for
Quaehirrh'rim. I also want to cram the nonstandard 'zu' row onto
there, in honour of Zecora's native language.)