+31w(176/140): add tentative numerals.csv to wordlists/
The exact particles/pronunciation are subject to revision, but the
basic logic of hexadecimal counting, sacred numbers '3', '5', '6',
'7', and 'twice/thrice' and 'less' for constructing the remaining
numbers seems sound. Decimal counting register (+ Arabic digits) for
reverse-engineered Earth technology is also a likely phenomenon. But
the system isn't finished until I've {recovered/assembled} the HirShi
numerals and bit fractions (1/2, 1/12, 1/16) used in high finance.

The numerous ways to pronounce 'bit' etc... are just going to get more
and more complicated. Calibrating the monetary/emotional balance of a
transaction is just not a concept in the present-day West.