+10w(245/210): catchup quota in wordlists/, init Elements theory
With the hirin'rin names of the Elements, we might say we have a full
spectrum of philosophical concepts. Word-colour synaesthesia in
Quaehirrh'rim is a whole other can of etymological worms to re-verify,
and it is indeed relevant to transmit accurately: the colour-value
of a word hugely influences naming as parents tend to pick names
that match the foal's natural colours.