+30w(305/245): catchup quota in wordlists/
Pace of transmission is picking up a bit. I anticipate a point at
which the construction of compound words explodes even further, at
which point the focus shifts to auguring the exact meaning of words
that *must* exist, as opposed to casting nets for words that *might*
exist. But I'd like to reach a balanced set of base morphemes covering
both the full range of meanings (comparable to some vocab set like
Basic English or a 日本語 Core 1000) and the full range of KinSyl
glyphs (perhaps ten per initial glyph, requiring a 640-word base).