removing show-translations/ in favour of separate repo
On reflection, it may be wise to use a separate Git repo for the MLP
show translations, in order to keep 'original' and 'derivative' material
apart. Distribution of the show scripts e.g. on wiki does not seem to be
objectionable to the current order of things, which I find
interesting, but it's still best to be cautious....