hirshi/: rework format for HirShi.csv, reached ?52/?1000
{the count will bifurcate into HirShi and KanShi categories,
aiming for ~1000 in both, but combining for now}

In light of committing the HirShi sketches to the repo not being a
realistic option, the CSV documents structural relations and provides
IDs that might be used to refer to an external document (to be shared
later).

Main augury source is, of course, qatei'yoshi documented on the show.
{Need to update the table with citations for which HirShi have been
adapted from a canonical qatei'yoshi.}

In practice converted both to a line style (for efficient writing) and
an illuminated-manuscript gradient style more closely mimicking the
naturally-occurring forms.